Change inline-block bidi tests to mitigate color blending
The fast/text/international/inline-block-with-mixed-direction-words.html
and fast/text/international/text-overflow-ellipsis-bidi.html fail in the
virtual antialiasedtext suite as left edge of the first character blends
slightly differently with the background due to antialiasing. By setting
a solid colored border in both the test and reference implementation
this is mitigated.

R=dglazkov@chromium.org
BUG=378610

Review URL: https://codereview.chromium.org/348143003

git-svn-id: svn://svn.chromium.org/blink/trunk@176667 bbb929c8-8fbe-4397-9dbb-9b2b20218538