Don't ever call mapRectToPaintInvalidationBacking past the repaint container.
In particular, don't call mapRectToPaintInvalidationBacking all the way up to
the transformed ancestor in the case of squashing.

Instead, call it up to the "repaint container" only (which in the case of squashing
is the *squashed* layer).  Then do the rest of the transformation with direct
arithmetic.

It turns out mapRectToPaintInvalidationBacking has several side effects that we do
not want.

BUG=386228

Review URL: https://codereview.chromium.org/342373003

git-svn-id: svn://svn.chromium.org/blink/trunk@176673 bbb929c8-8fbe-4397-9dbb-9b2b20218538