Rip out CompositingClipRects
We don't use this slot in the cache anymore. We added CompositingClipRects
early on in the compositing work, but never fully worked out all the bugs. Now
that we've improved the rest of the system, these rects don't seems as
necessary to cache for performance.

Review URL: https://codereview.chromium.org/339523007

git-svn-id: svn://svn.chromium.org/blink/trunk@176792 bbb929c8-8fbe-4397-9dbb-9b2b20218538