Mediastream: Simplify MediaStreamTrackSourcesRequestImpl class.
This CL removes m_scheduledEventTimer and m_protect, and use
ExecutionContext::postTask and WTF::bind.

We used m_protect to store a strong reference to |this| during the timer was
working. We need to use Persistent<> data member in Oilpan, but
we'd like to avoid it.

In this case, we don't need to use a Timer at all, and use
ExecutionContext::postTask and WTF::bind. WTF::bind automatically takes care of
lifetime of |this|.

TEST=none; No behavior changes.
BUG=

Review URL: https://codereview.chromium.org/346993003

git-svn-id: svn://svn.chromium.org/blink/trunk@176793 bbb929c8-8fbe-4397-9dbb-9b2b20218538