Update WebAudio mp3 test results
The platform-specific linux, mac, and win results are removed; testing
shows that the same file is produced on each of these platforms.

The actual reference file is updated due to the change in CL
https://codereview.chromium.org/350653004 which resulted in shorter decoded audio.

These tests aren't normally run on the trybots because ffmpeg is not built on the bots. These need to be run locally on each platform with a build that includes ffmpeg. I tested these manually on each platform.

BUG=135291
TBR=dalecurtis@chromium.org

Review URL: https://codereview.chromium.org/353763006

git-svn-id: svn://svn.chromium.org/blink/trunk@177000 bbb929c8-8fbe-4397-9dbb-9b2b20218538