Change port to use apache by default on win7, update test expectations.
This switches the default on Win 7 to use the native Win32 version of Apache
2.2 instead of LigHTTPd.  In theory Apache will substantially more stable
and also probably faster. I am hoping that this fixes the situation in bug
378444 where the server just doesn't work on some runs as well.

We also match more of the test expectations on other platforms.

If this change sticks and does improve flakiness, we'll try switching XP
as well in a later change.

This change also deletes all of the FlakyTest entries for http tests
on Win7, so that we can better get a real sense of the flakiness.

R=ojan@chromium.org
BUG=378444, 30536

Review URL: https://codereview.chromium.org/327423008

git-svn-id: svn://svn.chromium.org/blink/trunk@177046 bbb929c8-8fbe-4397-9dbb-9b2b20218538