Drop [TypeChecking=Nullable]; deduce from attribute type instead
[TypeChecking=Nullable] on a nullable attribute meant that the C++ getter
wasn't called with a "bool& isNull" argument, instead the C++ value in a
boolean context was used to determine nullness.

This was only used for attributes where the C++ value was a pointer, and
made perfect sense for such attributes. So much sense, in fact, that it
makes sense to let that be the default behavior for such attributes.

Review URL: https://codereview.chromium.org/352643005

git-svn-id: svn://svn.chromium.org/blink/trunk@177126 bbb929c8-8fbe-4397-9dbb-9b2b20218538