Remove bogus ASSERT in EventHandler::handleGestureEvent
I wrote this in r177164 thinking just that I wanted to verify the hit
test traversed all iframe boundaries, but of course there are other
widgets that aren't iframes (eg. plugins).

I tried changing the ASSERT to verify explicitly that the target
element is not an HTMLIFrameElement, but this doesn't work
either because touch adjustment may adjust the target to be over
an iframe element without necessarily hit-testing down into the
iframe.  This seems like we might want to change this to hit-test
down through the frame after adjusting to it, but the behavior has
been like this for a long time and we haven't seen an issue.

So just remove the ASSERT completely.  I don't think there's
anything meaningful to verify here.

TBRing to get the bots green ASAP and since the fix is trivial.

TBR=zeeshanq
BUG=390032

Review URL: https://codereview.chromium.org/359253002

git-svn-id: svn://svn.chromium.org/blink/trunk@177176 bbb929c8-8fbe-4397-9dbb-9b2b20218538