Fix RefPtr's move assignment operator
Fix RefPtr's move assignment operator so that the moved RefPtr is always
null after the assignment. Because the previous implementation was not
swapping with a temporary variable, this wasn't necessary the case and
this behavior is unexpected.

Thanks to Mikhail Pozdnyakov for pointing this out.

Review URL: https://codereview.chromium.org/354143005

git-svn-id: svn://svn.chromium.org/blink/trunk@177206 bbb929c8-8fbe-4397-9dbb-9b2b20218538