Update an Element's ancestor chain at the end of its activation
In the course of the an Element being activated, there is a possiblity
of the Element loosing its renderer (display: none being set on :active).
At the end of the activation, the element and its ancestor chain need to
be updated, so that the styles get reapplied. The current code skips the
last part if the active element did not have a renderer.

BUG=139660
TEST=fast/css/clear-activechain-list.html

Review URL: https://codereview.chromium.org/317733004

git-svn-id: svn://svn.chromium.org/blink/trunk@177215 bbb929c8-8fbe-4397-9dbb-9b2b20218538