[CSS Grid Layout] Update grid-auto-flow to the new syntax
In last versions of the spec grid-auto-flow syntax has changed. New
syntax is:
[ row | column ] && dense? | stack && [ row | column ]?

Implemented parsing for new syntax and added/modified test cases in
current layout tests.

Adapted code to keep the current behavior using "stack" as if it was
"none" in the previous syntax. This should be fixed in a follow-up patch
once "stack" is implemented.

Also "dense" needs to be reviewed. Right now auto-placement algorithm is
always "dense" and never "sparse".

BUG=384099

Review URL: https://codereview.chromium.org/333563003

git-svn-id: svn://svn.chromium.org/blink/trunk@177258 bbb929c8-8fbe-4397-9dbb-9b2b20218538