Adding a performance test for font fallback
Using a set of ~19000 characters extracted from DRT's font configuration, the
test is slicing them into smaller sets and rendering them in a page.  This
exercises WebFontInfo::fallbackFontForChar which is the cause for slowness,
as described in bug 266214.
This test case is preparation for landing the an optimization patch which
drastically reduces the number of fontconfig calls in the sandbox IPC process.

BUG=266214
R=eseidel,eae

Review URL: https://codereview.chromium.org/360363002

git-svn-id: svn://svn.chromium.org/blink/trunk@177337 bbb929c8-8fbe-4397-9dbb-9b2b20218538