Making it possible to import mediastream tests.
Since the W3C WebRTC tests involve permission prompts, I introduce a 
new mechanism whereby the test can specify class="instructions" 
for all elements that contain instructions to manual testers. 
These elements will be hidden by the importer upon import so 
they don't cause the test runner to be confused (it only wants 
to see pure testharness.js output).

Furthermore, since these tests depend on vendor-prefix.js to handle vendor 
prefixes, I add a slightly modified copy of that which is always on 
(i.e. doesn't need ?usePrefixes=1 in the URL to the test page), and logic 
to rewrite the paths so it points to our copy.

Note: must make the --use-fake-device-for-media-stream 
and --use-fake-ui-for-media-stream flags always be on for
layout tests before landing this patch.

BUG=389134

Review URL: https://codereview.chromium.org/353983003

git-svn-id: svn://svn.chromium.org/blink/trunk@177477 bbb929c8-8fbe-4397-9dbb-9b2b20218538